Non-persistent models, data annotations removed
