--- Models/AuthenticationModels/ForgotPasswordRequest.cs
@@ -1,9 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace LethalCompany_Backend.Models.AuthenticationModels;
-
-public class ForgotPasswordRequest
-{
-    [MaxLength(64), Required]
-    public string Email { get; set; }
-}
