+++ NonPersistentModels/AuthenticationModels/EmailExistsRequest.cs
@@ -1,0 +1,6 @@
+﻿namespace LethalCompany_Backend.Models.AuthenticationModels;
+
+public class EmailExistsRequest
+{
+    public string Email { get; set; }
+}
