+++ NonPersistentModels/AuthenticationModels/ForgotPasswordRequest.cs
@@ -1,0 +1,6 @@
+﻿namespace LethalCompany_Backend.Models.AuthenticationModels;
+
+public class ForgotPasswordRequest
+{
+    public string Email { get; set; }
+}
