+++ NonPersistentModels/AuthenticationModels/ForgotPasswordResponse.cs
@@ -1,15 +1,10 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace LethalCompany_Backend.Models.AuthenticationModels;
+﻿namespace LethalCompany_Backend.Models.AuthenticationModels;
 
 public class ForgotPasswordResponse
 {
-    [Required]
     public bool isSuccess { get; set; }
 
-    [MaxLength(256), Required]
     public string Token { get; set; }
 
-    [Required]
     public DateTime TokenExpireDate { get; set; }
 }
