+++ NonPersistentModels/AuthenticationModels/SetPasswordRequest.cs
@@ -1,0 +1,6 @@
+﻿namespace LethalCompany_Backend.Models.AuthenticationModels;
+
+public class SetPasswordRequest
+{
+    public string Password { get; set; }
+}
