+++ NonPersistentModels/MailSenderModels/SendEmailRequest.cs
@@ -1,6 +1,4 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace LethalCompany_Backend.Models.MailSenderModel;
+﻿namespace LethalCompany_Backend.Models.MailSenderModel;
 
 public enum MailType
 {
@@ -10,9 +8,7 @@ public enum MailType
 
 public class SendEmailRequest
 {
-    [MaxLength(64), Required]
     public string ReceiverMail { get; set; }
 
-    [Required]
     public MailType MailType { get; set; }
 }
