+++ NonPersistentModels/TokenModels/GenerateTokenResponse.cs
@@ -1,0 +1,8 @@
+﻿namespace LethalCompany_Backend.Models.TokenModels;
+
+public class GenerateTokenResponse
+{
+    public string Token { get; set; }
+
+    public DateTime TokenExpireDate { get; set; }
+}
